Перевел тесты на TS + добавил asyncTick для ожидания изменения в реактивности https://online.sbis.ru/opendoc.html?guid=58dbce83-3391-465f-95a2-091d15de9d8a
